add math functions on 5.3-pre
