+++ 5.1.d.ts
@@ -1,2 +1,4 @@
 import './core';
 import './special/5.1-only';
+import './special/5.3-pre';
+
