+++ 5.2.d.ts
@@ -1,3 +1,4 @@
 import './core';
 import './special/5.2-plus';
 import './special/5.2-only';
+import './special/5.3-pre';
