+++ core/math.d.ts
@@ -27,11 +27,6 @@ declare namespace math {
    */
   function atan(y: number, x?: number): number;
 
-  /**
-   * Returns the arc tangent of y/x (in radians), but uses the signs of both parameters to find the quadrant of the result. (It also handles correctly the case of x being zero.)
-   */
-  function atan2(y: number, x: number): number;
-
   /**   * Returns the smallest integral value larger than or equal to x.
    */
   function ceil(x: number): number;
@@ -40,11 +35,6 @@ declare namespace math {
    * Returns the cosine of x (assumed to be in radians).
    */
   function cos(x: number): number;
-
-  /**
-   * Returns the hyperbolic cosine of x.
-   */
-  function cosh(x: number): number;
 
   /**
    * Converts the angle x from radians to degrees.
@@ -67,29 +57,14 @@ declare namespace math {
   function fmod(x: number, y: number): number;
 
   /**
-   * Returns m and e such that x = m2e, e is an integer and the absolute value of m is in the range [0.5, 1) (or zero when x is zero).
-   */
-  function frexp(x: number): number;
-
-  /**
    * The float value HUGE_VAL, a value larger than any other numeric value.
    */
   const huge: number;
 
   /**
-   * Returns m2e (e should be an integer).
-   */
-  function ldexp(m: number, e: number): number;
-
-  /**
    * Returns the logarithm of x in the given base. The default for base is e (so that the function returns the natural logarithm of x).
    */
   function log(x: number, base?: number): number;
-
-  /**
-   * Returns the base-10 logarithm of x.
-   */
-  function log10(x: number): number;
 
   /**
    * Returns the argument with the maximum value, according to the Lua operator <. (integer/float)
@@ -112,35 +87,14 @@ declare namespace math {
   const pi: number;
 
   /**
-   * Returns x ^ y
-   */
-  function pow(x: number, y: number): number;
-
-  /**
    * Converts the angle x from degrees to radians.
    */
   function rad(x: number): number;
 
   /**
-   * This function is an interface to the simple pseudo-random generator function rand provided by ANSI C. (No guarantees can be given for its statistical properties.)
-   * When called without arguments, returns a uniform pseudo-random real number in the range [0,1). When called with an integer number m, math.random returns a uniform pseudo-random integer in the range [1, m]. When called with two integer numbers m and n, math.random returns a uniform pseudo-random integer in the range [m, n].
-   */
-  function random(m?: number, n?: number): number;
-
-  /**
-   * Sets x as the "seed" for the pseudo-random generator: equal seeds produce equal sequences of numbers.
-   */
-  function randomseed(x: number): number;
-
-  /**
    * Returns the sine of x (assumed to be in radians).
    */
   function sin(x: number): number;
-
-  /**
-   * Returns the hyperbolic sine of x.
-   */
-  function sinh(x: number): number;
 
   /**
    * Returns the square root of x. (You can also use the expression x^0.5 to compute this value.)
@@ -151,11 +105,6 @@ declare namespace math {
    * Returns the tangent of x (assumed to be in radians).
    */
   function tan(x: number): number;
-
-  /**
-   * Returns the hyperbolic tangent of x.
-   */
-  function tanh(x: number): number;
 
   /**
    * If the value x is convertible to an integer, returns that integer. Otherwise, returns nil.
